(#161) - fix viewCleanup 'ddoc/view' dedup
There was a problem before with the _local/mrviews
doc, where we weren't putting the fullViewName
(e.g. 'myddoc/myview'), and instead might put
just the abbreviated name ('myview').  This creates
duplication problems, but it's also something
that viewCleanup() didn't expect, hence errors
in that function.

The fix has an ignored if statement, because that
part is tested in a migration test in pouchdb
proper.