(#269) - allow custom query/viewCleanup implementations
Something that's come up while I was writing
[socket-pouch](https://github.com/nolanlawson/socket-pouch)
is that I need a way to tell mapreduce to not do the whole
dependentDb/separate database stuff. Right now it is designed
to create a secondary index in a new database, as long as it's
not an http adapter.

You can see why this is a problem for an adapter that needs to proxy
to a PouchDB in a remote server. Basically I want to disable the
query/viewCleanup stuff entirely and let the adapter handle it.

I think my solution is pretty elegant. Adapters can define an
`api._query` and `api._viewCleanup` and then the mapreduce plugin
will defer to that. If we wanted to, we could even define these functions
in `http.js` instead of mapreduce.